+++ src/vite-env.d.ts
@@ -1,5 +1,5 @@
 /// <reference types="svelte" />
 /// <reference types="vite/client" />
-/// <reference types="vite-plugin-pwa/client" />
+/// <reference types="@serwist/vite/typings" />
 
 declare const __APP_VERSION__: string;